@@ -1,22 +1,36 @@
 package net.skeagle.vrnenchants.enchant.enchantments;
 
-import net.minecraft.server.v1_16_R3.*;
-import net.skeagle.vrnenchants.VRNEnchants;
+import net.minecraft.ChatFormatting;
+import net.minecraft.network.chat.TextComponent;
+import net.minecraft.network.protocol.game.ClientboundAddMobPacket;
+import net.minecraft.network.protocol.game.ClientboundRemoveEntityPacket;
+import net.minecraft.network.protocol.game.ClientboundSetEntityDataPacket;
+import net.minecraft.network.protocol.game.ClientboundSetPlayerTeamPacket;
+import net.minecraft.network.syncher.EntityDataAccessor;
+import net.minecraft.network.syncher.EntityDataSerializers;
+import net.minecraft.network.syncher.SynchedEntityData;
+import net.minecraft.world.entity.EntityType;
+import net.minecraft.world.entity.monster.Shulker;
+import net.minecraft.world.scores.PlayerTeam;
+import net.minecraft.world.scores.Scoreboard;
+import net.minecraft.world.scores.Team;
 import net.skeagle.vrnenchants.enchant.*;
-import org.bukkit.Bukkit;
+import net.skeagle.vrnlib.misc.Task;
+import org.bukkit.ChatColor;
 import org.bukkit.Location;
 import org.bukkit.Material;
 import org.bukkit.block.Block;
-import org.bukkit.craftbukkit.v1_16_R3.CraftWorld;
-import org.bukkit.craftbukkit.v1_16_R3.entity.CraftPlayer;
+import org.bukkit.craftbukkit.v1_17_R1.CraftWorld;
+import org.bukkit.craftbukkit.v1_17_R1.entity.CraftPlayer;
 import org.bukkit.enchantments.Enchantment;
 import org.bukkit.entity.Player;
 import org.bukkit.event.block.Action;
 import org.bukkit.event.player.PlayerInteractEvent;
 
 import java.util.ArrayList;
-import java.util.Collections;
 import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
 
 @EnchDescription("Right clicking your pickaxe will allow you to see nearby ores for a few seconds.")
 public class EnchMineSight extends BaseEnchant implements ICooldown {
@@ -29,21 +43,21 @@ public class EnchMineSight extends BaseEnchant implements ICooldown {
         setCooldownMessage("&aYou are now able to use mine sight again.");
     }
 
-    private final HashMap<Block, EntityShulker> blockCorrespondingEntity = new HashMap<>();
-    private final ArrayList<EnumChatFormat> colorsUsed = new ArrayList<>();
+    private final Map<Block, Shulker> blockCorrespondingEntity = new HashMap<>();
+    private final List<ChatColor> colorsUsed = new ArrayList<>();
 
     @Override
     protected void onInteract(int level, PlayerInteractEvent e) {
         if (e.getAction() != Action.RIGHT_CLICK_BLOCK && e.getAction() != Action.RIGHT_CLICK_AIR) return;
         Player p = e.getPlayer();
-        ArrayList<Block> blocks = getBlocks(p.getLocation().getBlock(), level);
+        List<Block> blocks = getBlocks(p.getLocation().getBlock(), level);
         outline(p, blocks, level);
         setCooldown(p);
     }
 
-    private ArrayList<Block> getBlocks(Block start, int level) {
+    private List<Block> getBlocks(Block start, int level) {
         int radius = (level * 2) + 3;
-        ArrayList<Block> blocks = new ArrayList<>();
+        List<Block> blocks = new ArrayList<>();
         for (double x = start.getLocation().getX() - radius; x <= start.getLocation().getX() + radius; x++)
             for (double y = start.getLocation().getY() - radius; y <= start.getLocation().getY() + radius; y++)
                 for (double z = start.getLocation().getZ() - radius; z <= start.getLocation().getZ() + radius; z++)
@@ -51,33 +65,31 @@ public class EnchMineSight extends BaseEnchant implements ICooldown {
         return filterBlocks(blocks);
     }
 
-    private void outline(Player p, ArrayList<Block> blocks, int duration) {
+    private void outline(Player p, List<Block> blocks, int duration) {
         for (Block b : blocks) {
-            Location loc = b.getLocation();
-            WorldServer world = ((CraftWorld) p.getWorld()).getHandle();
-            EntityShulker shulk = new EntityShulker(EntityTypes.SHULKER, world);
-            shulk.setPositionRotation(loc.getX() + 0.5D, loc.getY(), loc.getZ() + 0.5D, 0, 0);
-            shulk.setHeadRotation(0);
+            Shulker shulk = new Shulker(EntityType.SHULKER, ((CraftWorld) b.getWorld()).getHandle());
+            shulk.setInvisible(true);
+            shulk.setXRot(0f);
+            shulk.setYRot(0f);
+            shulk.setPos(b.getLocation().getX() + 0.5, b.getLocation().getY(), b.getLocation().getZ() + 0.5);
             shulk.setInvulnerable(true);
-            shulk.setNoAI(true);
+            shulk.setNoAi(true);
             shulk.setSilent(true);
-            DataWatcher watcher = shulk.getDataWatcher();
-            watcher.set(new DataWatcherObject<>(0, DataWatcherRegistry.a), (byte) 96);
-            PacketPlayOutSpawnEntityLiving spawnEntityLiving = new PacketPlayOutSpawnEntityLiving(shulk);
-            ((CraftPlayer) p).getHandle().playerConnection.sendPacket(spawnEntityLiving);
-            PacketPlayOutEntityMetadata entityMetadata = new PacketPlayOutEntityMetadata(shulk.getId(), watcher, false);
-            ((CraftPlayer) p).getHandle().playerConnection.sendPacket(entityMetadata);
+            SynchedEntityData data = shulk.getEntityData();
+            data.set(new EntityDataAccessor<>(0, EntityDataSerializers.BYTE), (byte) 96);
+            ((CraftPlayer) p).getHandle().connection.send(new ClientboundAddMobPacket(shulk));
+            ((CraftPlayer) p).getHandle().connection.send(new ClientboundSetEntityDataPacket(shulk.getId(), data, false));
             blockCorrespondingEntity.put(b, shulk);
         }
-        for (EnumChatFormat color : colorsUsed)
+        for (ChatColor color : colorsUsed)
             setTeams(p, color);
-        Bukkit.getScheduler().runTaskLater(VRNEnchants.getInstance(), () -> {
-            PacketPlayOutEntityDestroy destroy;
-            for (EntityShulker shulk : blockCorrespondingEntity.values()) {
-                destroy = new PacketPlayOutEntityDestroy(shulk.getId());
-                ((CraftPlayer) p).getHandle().playerConnection.sendPacket(destroy);
+        Task.syncDelayed(() -> {
+            ClientboundRemoveEntityPacket destroy;
+            for (Shulker shulk : blockCorrespondingEntity.values()) {
+                destroy = new ClientboundRemoveEntityPacket(shulk.getId());
+                ((CraftPlayer) p).getHandle().connection.send(destroy);
             }
-        }, (long) (duration * 1.25) * 20);
+        },(long) (duration * 1.25) * 20);
     }
 
     @Override
@@ -86,22 +98,23 @@ public class EnchMineSight extends BaseEnchant implements ICooldown {
     }
 
     private enum Sorter {
-        COAL(EnumChatFormat.BLACK, Material.COAL_ORE),
-        IRON(EnumChatFormat.GRAY, Material.IRON_ORE),
-        GOLD(EnumChatFormat.GOLD, Material.GOLD_ORE),
-        LAPIS(EnumChatFormat.BLUE, Material.LAPIS_ORE),
-        REDSTONE(EnumChatFormat.RED, Material.REDSTONE_ORE),
-        DIAMOND(EnumChatFormat.AQUA, Material.DIAMOND_ORE),
-        EMERALD(EnumChatFormat.DARK_GREEN, Material.EMERALD_ORE),
-        QUARTZ(EnumChatFormat.WHITE, Material.NETHER_QUARTZ_ORE),
-        NETHERITE(EnumChatFormat.DARK_RED, Material.ANCIENT_DEBRIS),
-        NETHER_GOLD(EnumChatFormat.YELLOW, Material.NETHER_GOLD_ORE),
-        GILDED_BLACKSTONE(EnumChatFormat.YELLOW, Material.GILDED_BLACKSTONE);
+        COAL(ChatColor.BLACK, Material.COAL_ORE),
+        IRON(ChatColor.GRAY, Material.IRON_ORE),
+        GOLD(ChatColor.YELLOW, Material.GOLD_ORE),
+        LAPIS(ChatColor.BLUE, Material.LAPIS_ORE),
+        REDSTONE(ChatColor.RED, Material.REDSTONE_ORE),
+        DIAMOND(ChatColor.AQUA, Material.DIAMOND_ORE),
+        EMERALD(ChatColor.DARK_GREEN, Material.EMERALD_ORE),
+        QUARTZ(ChatColor.WHITE, Material.NETHER_QUARTZ_ORE),
+        NETHERITE(ChatColor.DARK_RED, Material.ANCIENT_DEBRIS),
+        NETHER_GOLD(ChatColor.YELLOW, Material.NETHER_GOLD_ORE),
+        GILDED_BLACKSTONE(ChatColor.YELLOW, Material.GILDED_BLACKSTONE),
+        COPPER(ChatColor.GOLD, Material.COPPER_ORE);
 
-        private final EnumChatFormat color;
+        private final ChatColor color;
         private final Material block;
 
-        Sorter(EnumChatFormat color, Material block) {
+        Sorter(ChatColor color, Material block) {
             this.color = color;
             this.block = block;
         }
@@ -110,31 +123,32 @@ public class EnchMineSight extends BaseEnchant implements ICooldown {
             return block;
         }
 
-        public static EnumChatFormat getColorFrom(Block b) {
+        public static ChatColor getColorFrom(Block b) {
             for (Sorter sorter : Sorter.values())
                 if (b.getType() == sorter.block)
                     return sorter.color;
-            return EnumChatFormat.WHITE;
+            return ChatColor.WHITE;
         }
     }
 
-    private void setTeams(Player p, EnumChatFormat color) {
+    private void setTeams(Player p, ChatColor color) {
         Scoreboard board = new Scoreboard();
-        ScoreboardTeam team = new ScoreboardTeam(board, "MINESIGHT@@" + color.character);
-        team.setPrefix(new ChatMessage("§" + color.character));
-        team.setNameTagVisibility(ScoreboardTeamBase.EnumNameTagVisibility.ALWAYS);
-        team.setCollisionRule(ScoreboardTeamBase.EnumTeamPush.ALWAYS);
-        team.setDisplayName(new ChatMessage("minesight" + color.character));
-        team.setColor(color);
-        ((CraftPlayer) p).getHandle().playerConnection.sendPacket(new PacketPlayOutScoreboardTeam(team, 0));
-        for (Block b : blockCorrespondingEntity.keySet())
+        String s = "MINESIGHT@@" + color.getChar();
+        PlayerTeam nmsTeam = board.addPlayerTeam(s);
+        nmsTeam.setCollisionRule(Team.CollisionRule.NEVER);
+        nmsTeam.setDisplayName(new TextComponent("minesight" + color.getChar()));
+        nmsTeam.setColor(ChatFormatting.getByCode(color.getChar()));
+        System.out.println(nmsTeam.getColor());
+        ((CraftPlayer) p).getHandle().connection.send(ClientboundSetPlayerTeamPacket.createAddOrModifyPacket(nmsTeam, true));
+        for (Block b : blockCorrespondingEntity.keySet()) {
             if (Sorter.getColorFrom(b) == color)
-                ((CraftPlayer) p).getHandle().playerConnection.sendPacket(
-                        new PacketPlayOutScoreboardTeam(team, Collections.singletonList(String.valueOf(blockCorrespondingEntity.get(b).getUniqueID())), 3));
+                ((CraftPlayer) p).getHandle().connection.send(ClientboundSetPlayerTeamPacket.createPlayerPacket(nmsTeam,
+                        blockCorrespondingEntity.get(b).getStringUUID(), ClientboundSetPlayerTeamPacket.Action.ADD));
+        }
     }
 
-    private ArrayList<Block> filterBlocks(ArrayList<Block> blocks) {
-        ArrayList<Block> newBlocks = new ArrayList<>();
+    private List<Block> filterBlocks(List<Block> blocks) {
+        List<Block> newBlocks = new ArrayList<>();
         for (Block b : blocks)
             for (Sorter sort : Sorter.values())
                 if (b.getType() == sort.getBlock()) {
